add tokenator crate
also remove a lot of the "advanced" token parsing style
which was a bit too verbose for my tastes
