+++ Cargo.toml
@@ -5,7 +5,7 @@ members = [
     "crates/notedeck_chrome",
     "crates/notedeck_columns",
 
-    "crates/enostr",
+    "crates/enostr", "crates/tokenator",
 ]
 
 [workspace.dependencies]
@@ -34,6 +34,7 @@ nostrdb = { git = "https://github.com/damus-io/nostrdb-rs", rev = "2111948b078b2
 notedeck = { path = "crates/notedeck" }
 notedeck_chrome = { path = "crates/notedeck_chrome" }
 notedeck_columns = { path = "crates/notedeck_columns" }
+tokenator = { path = "crates/tokenator" }
 open = "5.3.0"
 poll-promise = { version = "0.3.0", features = ["tokio"] }
 puffin = { git = "https://github.com/jb55/puffin", package = "puffin", rev = "70ff86d5503815219b01a009afd3669b7903a057" }
