+++ crates/notedeck_columns/src/storage/decks.rs
@@ -18,6 +18,7 @@ use crate::{
 };
 
 use notedeck::{storage, DataPath, DataPathType, Directory};
+use tokenator::{ParseError, TokenParser, TokenSerializable, TokenWriter};
 
 pub static DECKS_CACHE_FILE: &str = "decks_cache.json";
