+++ crates/notedeck_columns/src/storage/mod.rs
@@ -1,11 +1,5 @@
 mod decks;
 mod migration;
-mod token_parser;
 
 pub use decks::{load_decks_cache, save_decks_cache, DECKS_CACHE_FILE};
 pub use migration::{deserialize_columns, COLUMNS_FILE};
-
-pub use token_parser::{
-    ParseError, Payload, Token, TokenAlternatives, TokenParser, TokenPayload, TokenSerializable,
-    TokenWriter, UnexpectedToken,
-};
