+++ crates/tokenator/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "tokenator"
+version = "0.1.0"
+edition = "2021"
+description = "A simple library for parsing a serializing string tokens"
+
+[dependencies]
