+++ crates/tokenator/README.md
@@ -1,0 +1,5 @@
+
+# tokenator
+
+Tokenator is a simple string token parser and serializer.
+
